refact: rename method and file
